inject key to fix testing setup
