+++ src/tests/attestation.test.ts
@@ -1,11 +1,15 @@
 import "./setup"; // Must be imported first to set up environment variables
 import { describe, expect, it } from "bun:test";
 import { createApp } from "../app";
+import { deriveDevEncryptionKey } from "../key";
+import { initializeServices } from "../services";
 import { env } from "../config";
 
 describe("Attestation", () => {
 	it("should return 200 and a public key", async () => {
-		const app = await createApp();
+		const encryptionKey = await deriveDevEncryptionKey();
+		const services = await initializeServices(env, encryptionKey);
+		const app = await createApp(services);
 
 		const req = new Request("http://localhost/v1/attestation/public-key", {
 			headers: {
